requirements.txt
# Core API Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6  # Required for file uploads

# AI/ML Libraries
langchain>=0.1.0
langchain-openai>=0.0.5
transformers>=4.35.0
torch>=2.1.0

# Redis for state management
redis>=5.0.0

# File processing
pdfplumber>=0.10.0
pdf2image>=1.16.0
pytesseract>=0.3.10
Pillow>=10.0.0

# OCR Libraries - Hugging Face based
# transformers already included above
# torch already included above

# Data validation and serialization
pydantic>=2.5.0
pydantic-settings>=2.1.0

# YAML processing
PyYAML>=6.0.1

# HTTP client for callbacks
httpx>=0.25.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0

# Optional: Ollama client (if using local LLMs)
ollama>=0.1.0

# Logging and monitoring
structlog>=23.2.0

# Additional dependencies for OCR and image processing
numpy>=1.21.0
opencv-python>=4.5.0
datasets>=2.0.0  # For Hugging Face datasets if needed
